Fixed saving options form action
In `src/WordPressAdmin.php` the action is defined as following:
```
add_action(
    'admin_post_wp2static_ui_save_options',
    [ 'WP2Static\Controller', 'wp2staticUISaveOptions' ],
    10,
    0
);
```

However, in this form we try to call action `admin_post_wp2staticUISaveOptions`, when
in reality it should be `admin_post_wp2static_ui_save_options`. This results in WP admin
a problem when trying to save the settings.