fix plugin assets detection duplicating urls
I was experiencing an issue with duplicating site paths included in all detected plugin paths. To replicate this install and activate the Yoast plugin, and look at the detected urls. All plugin asset urls will not begin with "/wp-content..." and instead contain the entire site url. This is resulting in errors that are displayed during the crawl. 

Additionally this makes it so that the yoast sitemap stylesheet assets are not included, which after modifying the initial crawl list (thanks for including this filter!) the yoast sitemaps are crawled and missing the xsl styles that they reference.

This change alters the plugin asset detection to use nearly the same code to the theme asset detection.