Refactor code to separate detection and enqueueing
So that we can count without enqueueing while checking status and figure out if we need to actually detect and enqueue URLs due to there being new ones.

Also adds an option for doing a "quiet" count so that we don't write to CLI output when doing a detect for URL count.