Fix error when a sitemap path starts with //
When creating a Psr7 request with a base_uri set in the Client and a
path that starts with // (like '//wp-sitemap.xml'), the path is
interpreted as an absolute URL and we get a ClientException for
'unknown host: wp-sitemap.xml'. This is fixed by removing base_uri
from the client and always making Requests with the full url.

This was discovered while debugging #824 and may address that issue.